requirements.txt
Babel==2.6.0
dj-database-url==0.5.0
Django==2.1.2
django-countries==5.3.2
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-multiselectfield==0.1.8
django-phonenumber-field==2.1.0
django-phonenumbers==1.0.1
django-registration-redux==2.5
gunicorn==19.9.0
phonenumbers==8.9.16
Pillow==5.3.0
psycopg2==2.7.7
pytz==2018.7
virtualenv==16.0.0
whitenoise==4.1.2
django-currentuser
